ImageCache: TRY using boost::flat_map (and supporting changes)
Also a supporting change of defining an ordering to TileID, which also
goes a lot faster if we cache the hashed value in the struct.

But in the end, I just couldn't get faster renders with flat_map.  So
leave it disabed for now, with #define's to turn it on, in case other
changes make us want to revive it or re-benchmark it.