checkpoint
Changed the signature of the most general read_native_tiles to include
subimage and miplevel indices; it is now assumed to perform a
seek_subimage internally, and also to be fully thread-safe.

We were actually more clever than just putting a lock around the
whole thing; the method was refactored so any data currying happens
after the lock is released.