Overhaul ImageBufAlgo API
We make four large changes across the board in ImageBufAlgo.

1. No more raw pointers.

All functions that take per-channel constants, or destinations for
per-channel outputs, in the form of (pointer, length) -- or worse, just
a pointer and an assumed length big enough to hold all channels -- are
being deprecated, in favor of passing span<>. This will let the IBA
function implementations check lengths and deal appropriately, also
makes it very easy for the caller to pass a single float that will be
understood to be replicated for all channels.

2. ImageBuf return results

All existing functions of the form

    bool IBA::func (ImageBuf& result, ...input args...)

are getting an additional variant,

    ImageBuf IBA::func (...input args...);

For most ordinary uses, this transforms the awkward

    ImageBuf Result;
    ImageBufAlgo::add (Result, A, B);

into the more compact and intuitive

    ImageBuf Result = ImageBufAlgo::add (A, B);

Through the magic of C++11 "move" semantics (aka "rvalue assignment"),
this does not incur an unnecessary copy of the whole ImageBuf.

The old function forms will remain as first class citizens and are stil
useful in cases where you are repeatedly doing operations (so want to avoid
repeated allocation), or you want to use ROI's to write partial results of
the operation into a sub-region of an existing image.

This carries over to the Python API as well, where the old idiom

    Result = ImageBuf()
    ImageBufAlgo.add (Result, A, B)

was especially "non-Pythonic" looking, compared to

    Result = ImageBufAlgo.add (A, B)

3. Reduce combinatorics of Image vs constant.

For some functions, such as mad(), which computes A*B+C, a whole bunch
of clutter is necessary to handle the various combinations of

    image * image + image  ->  mad (const IB&, const IB&, const IB&)
    image * const + image  ->  mad (const IB&, span<const float>, const IB&)
    image * image + const  ->  mad (const IB&, const IB&, span<>)
    const * image + const      ...
    etc.

In fact, we need 7 variants (excluding const*const+const, which is not
useful because you don't know how big the resulting image would be).

These have all been replaced by a single variant,

    mad (Image_or_Const A, Image_or_Const B, Image_or_Const C)

using a special new class just used for parameter passing,
Image_or_Const, which can "wrap" either an ImageBuf or a float or a
span<const float>, and let the implementation sort it all out
internally.  It really de-clutters the declarations! And it also forces
us to ensure in these cases that all the combinations are accounted for.

4. Some IBA functions that generate structs with data, such as compare(),
computePixelStats(), and histogram(), now directly return those structures
instead of needing them pre-declared and passed as reference paramters.
Thanks to the magic of C++11 'move' semantics, this is no longer expensive,
and it's a more intuitive way to structure the API. (The old versions of
the functions still are available, but are considered deprecated and will
be removed in future releases.)

In the process of these changes, I've done a substantial rewrite of both
the IBA and Python chapters of the docs, as well as all the verbose
comments in imagebufalgo.h itself that explains the functions.