IBA::deep_holdout -- fix bug when channels didn't match
deep_holdout() was crashing when the holdout image had more channels
than the input image. It was improper to rely on IBAPrep's default behavior
of setting the result channels to the maximum of the channels in any of
the input images. In this case, we want to force it to match the "shape"
of just the first (source) input, not the second input that only serves
to act as the depth holdout.

While I was there, though, I tightened up the logic of how IBAPrep deals
with the force_spec parameter, and also made that parameter a const ptr,
as it always should have been.

deep_holdout fixes when src and holdout image resolutions don't match

Rename deep_holdout to deep_cull, do proper implementation of deep_holdout

Decided that deep_holdout should be a different behavior that is the
deep extension to a 2D holdout (holdout blocks the input image, but leaves
a hole).

Lots of little additions to DeepData along the way.

parse_roi() turns many string formats into ROIs

oiiotool --dumpdata:roi=...

New dumpdata modifier roi= lets you dump pixel-by-pixel data for just a
subset of the image, including just a single pixel if you want. For example,

    oiiotool -dumpdata:roi=659,660,393,394 buggy_deep_image.exr

It can accept either "geometry" style ("1x1+659+393") or "ROI style"
("659,660,393,394"). Note that the range style uses the range notation
where end values are exclusive, not inclusive.

holdout wip

Add Deepdata::is_tidy()

WIP experimental

WIP works

Remove debug scaffolding and ratholes

Deepdata::split_all, and change split to return an int

tests -- works for nuke, fails unit tests, ugh