Change span size to unsigned.
Alas, the C++ committee in its wisdom standardized on span<> for C++20
slighty differently than the draft we based our span on. Specifically,
they changed size() to be unsigned (like the other containers) and added
a separate ssize() to return a signed size.

Hopefully this won't be a major disruption.