ImageInput: Errors that cause open() to fail did not close the file handle.
Fixes #2240

This was reported as a problem with DDS, but after being urged to poke
around at all the ImageInput subclasses, it was widespread that errors
in the open() methods did not always release open file handles.

As a practical matter, this is not super important, because in all
cases, destroying the ImageInput -- expected to happen by the caller
if open fails -- would also cause all the handles to be closed. But I
agree that it's nicer to do so immediately upon realizing it's an
unreadable file.

I added more tests to imageinout_test, where for each format we create
a bogus file and check that the format's valid_file() fails properly,
and also check that valid_file() succeeds properly for a valid file.
(And, oops, in the process detected that SGI format did not work properly
because of failure to deal with endianness in valid_file, though it did
so for open.)

This is one of those times when you think you're pulling just one
little loose thread on the sweater, and next thing you know it exposes
a bunch of issues spread all over that you never knew were a
problem. But that's how software gets better.

Also fix broken sgi valid_file that was wrong because no endian swap.

And OpenVDB: get rid of unused field and clarify comments.
The file per se doesn't need to be held open beyond the open() file,
and we didn't. But we still had a m_input field that was never really
assigned to. Get rid of the cruft, and add a comment so the next
person in this code doesn't waste time wondering why the handle never
seems to be saved.