Add ssize to span
C++20 has channged span's size() to be unsigned (ugh, but matching
.size() for other types) and adds ssize() for when the client wants a
signed value.  We follow suit here, but we need to stage it: First,
add ssize(), which is back-compatible, and backport it to the last
release. Then later, at the right major release point, we will make
the compatibility-breakinng change of making size() be size_t.