parallel.h refactoring
Hide nearly all of the implementation of the parallel_for family of
methods rather than expose them as inline. This gives us more freedom
to change the implementation without breaking ABI.

Deprecate the parallel methods that take task functions whose first
parameter is the thread ID. The only reason they existed is because
our cobbled together thread_pool used that in its inner workings.  But
there are good reasons why we should not expose that:

* We never used it anyway.

* It was not very useful, since sometimes it was a real thread ID, but
  other times it was -1 in cases where the calling thread executed the
  task directly.

* It is inconsistent with other thread pool implementations that we may
  wish to try in the future.

So better just to not expose those thread IDs at all. Mark those
versions of the parallel loops as deprecated and schedule them for
removal in OIIO 3.0.