WIP bug fix for libheif
libheif 1.10, which was just released, seems to have some severe errors
writing, leaving corrupt files. At least for Homebrew?

First I thought maybe the bugs were in the C++ layer, so this commit
changes the writer to use the C API. Still broken.

So maybe it's not C++? Maybe it's using interleaved channels?